backend/requirements.txt
# 后端依赖
pydantic-settings>=2.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.3.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.26.0
google-generativeai>=0.8.0
python-alipay-sdk>=3.0.0
